+++ src/Themes/index.ts
@@ -1,0 +1,8 @@
+import Dynamite_High_Contrast from './Dynamite_High_Contrast.json';
+import Dynamite from './Dynamite.json';
+import { Theme } from '../constant.enum.modal';
+
+export default {
+  [Theme.DYNAMITE_HIGH_CONTRAST]: Dynamite_High_Contrast,
+  [Theme.DYNAMITE]: Dynamite
+}