+++ src/enum.ts
@@ -1,5 +1,6 @@
 export const ADD_COLOR = 'themeChanger.addColor';
 export const CLEAR_COLORS = 'themeChanger.clearColors';
+export const RESET_COLORS = 'themeChanger.resetColors';
 export const RESET_THEME = 'themeChanger.resetTheme';
 export const CHANGE_THEME_COLOR = 'themeChanger.changeThemeColor';
 export const COLORS_VIEW = 'themeChanger.sidebar-colorsView';