+++ src/extension.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { ADD_COLOR, CHANGE_THEME_COLOR, CLEAR_COLORS, RESET_THEME } from './enum';
+import { ADD_COLOR, CHANGE_THEME_COLOR, CLEAR_COLORS, RESET_COLORS, RESET_THEME } from './enum';
 import { SidebarColorsView } from './SidebarColorsView';
 import { StatusbarUi } from './StatusBarUI';
 import { ThemeChanger } from './theme-changer';
@@ -11,16 +11,17 @@ export function activate(context: vscode.ExtensionContext) {
 
   context.subscriptions.push(vscode.window.registerWebviewViewProvider(SidebarColorsView.viewType, sidebarColorsView));
 
-  // Change Theme Color
-  context.subscriptions.push(
-    vscode.commands.registerCommand(CHANGE_THEME_COLOR, () => sidebarColorsView.changeThemeColor())
-  );
+  // Change to next Theme Color
+  context.subscriptions.push(vscode.commands.registerCommand(CHANGE_THEME_COLOR, () => sidebarColorsView.changeThemeColor()));
 
   // Add Color
   context.subscriptions.push(vscode.commands.registerCommand(ADD_COLOR, () => sidebarColorsView.addColor()));
 
   // Clear Colors
   context.subscriptions.push(vscode.commands.registerCommand(CLEAR_COLORS, () => sidebarColorsView.clearColors()));
+
+  // Reset Colors
+  context.subscriptions.push(vscode.commands.registerCommand(RESET_COLORS, () => sidebarColorsView.resetColors()));
 
   // Rest to VsCode default theme. Clears all generated theme colors
   context.subscriptions.push(vscode.commands.registerCommand(RESET_THEME, () => sidebarColorsView.resetTheme()));
@@ -30,4 +31,4 @@ export function activate(context: vscode.ExtensionContext) {
 }
 
 // this method is called when your extension is deactivated
-export function deactivate() {}
+export function deactivate() { }